bot/pyproject.toml
[project]
name = "bot"
version = "0.1.0"
description = "LMS Telegram Bot"
dependencies = [
    "python-dotenv",
    "httpx",
    "aiogram",
]
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
